+++ R/leaflet_polar_pollroseMap.R
@@ -54,10 +54,8 @@ pollroseMap <- function(data,
                         key = FALSE,
                         draw.legend = TRUE,
                         collapse.control = FALSE,
-                        iconWidth = 200,
-                        iconHeight = 200,
-                        fig.width = 3.5,
-                        fig.height = 3.5,
+                        d.icon = 200,
+                        d.fig = 3.5,
                         type = NULL,
                         ...) {
   if (!is.null(type)) {
@@ -68,14 +66,13 @@ pollroseMap <- function(data,
   }
 
   # assume lat/lon
-  latlon <- assume_latlon(
-    data = data,
-    latitude = latitude,
-    longitude = longitude
-  )
+  latlon <- assume_latlon(data = data,
+                          latitude = latitude,
+                          longitude = longitude)
   latitude <- latlon$latitude
   longitude <- latlon$longitude
 
+  # prep data
   data <-
     prepMapData(
       data = data,
@@ -99,12 +96,6 @@ pollroseMap <- function(data,
     theBreaks <- 6
   }
 
-  # define plotting function
-  args <- list(...)
-  fun <- function(...) {
-    rlang::exec(openair::pollutionRose, alpha = alpha, annotate = FALSE, statistic = statistic, breaks = theBreaks, !!!args, ...)
-  }
-
   # identify splitting column (defaulting to pollutant)
   if (length(pollutant) > 1) {
     split_col <- "pollutant_name"
@@ -115,34 +106,45 @@ pollroseMap <- function(data,
     split_col <- "pollutant_name"
   }
 
-  # create icons
-  icons <-
-    data %>%
-    dplyr::group_split(.data[[split_col]]) %>%
-    rlang::set_names(levels(data[[split_col]])) %>%
-    purrr::imap(
-      .f = ~ create_icons(
-        data = .x, fun = fun, pollutant = "conc", split = .y,
-        lat = latitude, lon = longitude, x = x, cols = cols,
-        key = key, fig.width = fig.width, fig.height = fig.height,
-        iconWidth = iconWidth, iconHeight = iconHeight, ...
-      )
+  # define function
+  fun <- function(data) {
+    openair::pollutionRose(
+      data,
+      pollutant = "conc",
+      statistic = statistic,
+      breaks = theBreaks,
+      plot = FALSE,
+      cols = cols,
+      alpha = alpha,
+      key = key,
+      annotate = FALSE,
+      ...,
+      par.settings = list(axis.line = list(col = "transparent"))
     )
+  }
 
-  # plot leaflet
-  map <-
-    makeMap(
+  # create temp directory
+  tempdir <- tempdir()
+
+  # plot and save static markers
+  plots_df <-
+    create_static_markers(
+      fun = fun,
       data = data,
-      icons = icons,
-      provider = provider,
-      longitude = longitude,
+      dir = tempdir,
       latitude = latitude,
-      popup = popup,
-      label = label,
+      longitude = longitude,
       split_col = split_col,
-      collapse = collapse.control
+      d.fig = d.fig,
+      popup = popup,
+      label = label
     )
 
+  # create leaflet map
+  map <-
+    make_leaflet_map(plots_df, latitude, longitude, provider, d.icon, popup, label, split_col, collapse.control)
+
+  # add legend if breaks are defined
   if (!is.null(breaks) & draw.legend) {
     map <-
       leaflet::addLegend(
@@ -157,5 +159,6 @@ pollroseMap <- function(data,
       )
   }
 
-  map
+  # return map
+  return(map)
 }