+++ R/utils-map.R
@@ -84,7 +84,7 @@ checkMapPrep <-
       if (wd %in% Names & is.numeric(mydata[, wd])) {
         ## check for wd <0 or > 360
         if (any(sign(mydata[[wd]][!is.na(mydata[[wd]])]) == -1 |
-          mydata[[wd]][!is.na(mydata[[wd]])] > 360)) {
+                mydata[[wd]][!is.na(mydata[[wd]])] > 360)) {
           warning("Wind direction < 0 or > 360; removing these data")
           mydata[[wd]][mydata[[wd]] < 0] <- NA
           mydata[[wd]][mydata[[wd]] > 360] <- NA
@@ -205,199 +205,6 @@ prepMapData <- function(data, pollutant, control, ..., .to_narrow = TRUE) {
   return(data)
 }
 
-
-#' Save an openair plot as a temp image to use as an icon
-#' @noRd
-save_icon_image <-
-  function(data,
-           fun,
-           dir,
-           pollutant,
-           split,
-           lat,
-           lon,
-           cols,
-           key,
-           fig.width,
-           fig.height,
-           ...) {
-    id <- paste0(data[[lat]][1], data[[lon]][1])
-
-    grDevices::png(
-      type = "cairo-png",
-      filename = paste0(dir, "/", id, "_", split, ".png"),
-      width = fig.width * 300,
-      height = fig.height * 300,
-      res = 300,
-      bg = "transparent"
-    )
-
-    plt <- fun(
-      data,
-      pollutant = pollutant,
-      key = key,
-      cols = cols,
-      par.settings = list(axis.line = list(col = "transparent")),
-      ...
-    )
-
-    grDevices::dev.off()
-  }
-
-#' Save all openair plots as images and read as leaflet icons
-#' @noRd
-create_icons <-
-  function(data,
-           fun,
-           pollutant,
-           split,
-           lat,
-           lon,
-           cols,
-           key,
-           fig.width,
-           fig.height,
-           iconWidth,
-           iconHeight,
-           ...) {
-    # where to write files
-    icon_dir <- tempdir()
-
-    # drop missing data
-    data <- tidyr::drop_na(data, .data[[pollutant]])
-
-    # go through all sites and make some plot
-    data %>%
-      dplyr::arrange(.data[[lat]], .data[[lon]]) %>%
-      dplyr::group_split(.data[[lat]], .data[[lon]]) %>%
-      purrr::walk(
-        .f = ~ save_icon_image(
-          fun = fun,
-          dir = icon_dir,
-          pollutant = pollutant,
-          split = split,
-          lat = lat,
-          lon = lon,
-          cols = cols,
-          key = key,
-          fig.width = fig.width,
-          fig.height = fig.height,
-          ...
-        )
-      )
-
-    dat2 <- data %>%
-      dplyr::arrange(.data[[lat]], .data[[lon]]) %>%
-      dplyr::mutate(id = paste0(.data[[lat]], .data[[lon]]))
-
-    # definition of 'icons' aka the openair plots
-    leafIcons <-
-      lapply(
-        paste0(
-          icon_dir, "/", unique(dat2$id), "_", split, ".png"
-        ),
-        leaflet::makeIcon,
-        iconWidth = iconWidth,
-        iconHeight = iconHeight
-      )
-    names(leafIcons) <- unique(dat2$id)
-    class(leafIcons) <- "leaflet_icon_set"
-
-    leafIcons
-  }
-
-
-#' Make a leaflet map
-#' @noRd
-makeMap <-
-  function(data,
-           icons,
-           provider,
-           longitude,
-           latitude,
-           split_col,
-           popup,
-           label,
-           collapse) {
-    provider <- unique(provider)
-
-    # data for plotting
-    plot_data <-
-      data %>%
-      dplyr::group_by(.data[[latitude]], .data[[longitude]], .data[[split_col]]) %>%
-      dplyr::mutate(dc = mean(!is.na(.data[["conc"]]))) %>%
-      dplyr::ungroup() %>%
-      dplyr::distinct(.data[[latitude]], .data[[longitude]], .data[[split_col]], .keep_all = TRUE) %>%
-      dplyr::arrange(.data[[latitude]], .data[[longitude]])
-
-    # create leaflet map
-    m <- leaflet::leaflet()
-
-    # add tiles
-    for (j in seq(length(provider))) {
-      m <- leaflet::addProviderTiles(
-        map = m,
-        provider = provider[j],
-        group = provider[j]
-      )
-    }
-
-    # add markers
-    for (i in names(icons)) {
-      plot_data_i <-
-        dplyr::filter(plot_data, .data[[split_col]] == i) %>%
-        dplyr::filter(.data$dc != 0)
-
-      if (!is.null(popup)) {
-        thePopup <- plot_data_i[[popup]]
-      } else {
-        thePopup <- popup
-      }
-
-      if (!is.null(label)) {
-        theLabel <- plot_data_i[[label]]
-      } else {
-        theLabel <- label
-      }
-
-      # only plot markers where there is data
-      m <- leaflet::addMarkers(
-        m,
-        data = plot_data_i,
-        lng = plot_data_i[[longitude]],
-        lat = plot_data_i[[latitude]],
-        icon = icons[[i]],
-        popup = thePopup,
-        label = theLabel,
-        group = i %>% quickTextHTML()
-      )
-    }
-
-    # add layer control for pollutants/providers
-    if (length(icons) > 1 & length(provider) > 1) {
-      m <-
-        leaflet::addLayersControl(
-          m,
-          options = leaflet::layersControlOptions(collapsed = collapse),
-          baseGroups = names(icons) %>% purrr::map_vec(quickTextHTML),
-          overlayGroups = provider
-        )
-    } else if (length(icons) > 1 & length(provider) == 1) {
-      m <- leaflet::addLayersControl(m,
-        options = leaflet::layersControlOptions(collapsed = collapse),
-        baseGroups = names(icons) %>% purrr::map_vec(quickTextHTML)
-      )
-    } else if (length(provider) > 1 & length(icons) == 1) {
-      m <- leaflet::addLayersControl(m,
-        options = leaflet::layersControlOptions(collapsed = collapse),
-        baseGroups = provider
-      )
-    }
-
-    # return
-    return(m)
-  }
-
 #' guess latlon
 #' @noRd
 assume_latlon <- function(data, latitude, longitude) {
@@ -422,7 +229,7 @@ assume_latlon <- function(data, latitude, longitude) {
     len <- length(out)
     if (len > 1) {
       cli::cli_abort("Cannot identify {name}: Multiple possible matches ({out})",
-        call = NULL
+                     call = NULL
       )
       return(NULL)
     } else if (len == 0) {
@@ -435,7 +242,6 @@ assume_latlon <- function(data, latitude, longitude) {
   }
 
   if (is.null(latitude) | is.null(longitude)) {
-    cli::cli_h1("Assuming Latitude and/or Longitude")
     if (is.null(latitude)) {
       latitude <- guess_latlon(data, "lat")
     } else {
@@ -480,3 +286,201 @@ getBreaks <- function(breaks, ws.int, vec, polrose) {
   breaks <- sort(breaks)
   breaks
 }
+
+#' make leaflet map from scratch
+#' @noRd
+make_leaflet_map <-
+  function(data,
+           latitude,
+           longitude,
+           provider,
+           d.icon,
+           popup,
+           label,
+           split_col,
+           collapse.control) {
+    # create map
+    map <- leaflet::leaflet(data)
+
+    # add provider tiles
+    for (i in unique(provider)) {
+      map <- leaflet::addProviderTiles(map, i, group = i)
+    }
+
+    # add markers
+    marker_arg <- list(
+      map = map,
+      lat = data[[latitude]],
+      lng = data[[longitude]],
+      icon = leaflet::makeIcon(
+        iconUrl = data$url,
+        iconHeight = d.icon,
+        iconWidth = d.icon,
+        iconAnchorX = d.icon / 2,
+        iconAnchorY = d.icon / 2
+      ),
+      group = quickTextHTML(data[[split_col]])
+    )
+
+    if (!is.null(popup)) {
+      marker_arg <- append(marker_arg, list(popup = data[[popup]]))
+    }
+    if (!is.null(label)) {
+      marker_arg <- append(marker_arg, list(label = data[[label]]))
+    }
+
+    map <- rlang::exec(leaflet::addMarkers,!!!marker_arg)
+
+    # add layer control menu
+    flag_provider <- dplyr::n_distinct(provider) > 1
+    flag_split <- dplyr::n_distinct(data[[split_col]]) > 1
+    opts <- leaflet::layersControlOptions(collapsed = collapse.control, autoZIndex = FALSE)
+
+    if (flag_provider & flag_split) {
+      map <-
+        leaflet::addLayersControl(map,
+                                  baseGroups = quickTextHTML(unique(data[[split_col]])),
+                                  overlayGroups = provider, options = opts) %>%
+        leaflet::hideGroup(group = provider[-1])
+    } else if (flag_provider & !flag_split) {
+      map <- leaflet::addLayersControl(map, baseGroups = provider, options = opts) %>%
+        leaflet::hideGroup(group = provider[-1])
+    } else if (!flag_provider & flag_split) {
+      map <-
+        leaflet::addLayersControl(map, baseGroups = quickTextHTML(unique(data[[split_col]])), options = opts)
+    }
+
+    return(map)
+  }
+
+#' theme for static maps
+#' @noRd
+theme_static <- function() {
+  ggplot2::`%+replace%`(
+    ggplot2::theme_minimal(),
+    ggplot2::theme(panel.border = ggplot2::element_rect(fill = NA, color = "black"))
+  )
+}
+
+#' Create markers for the static plots
+#' @param fun function of "data" to create plot
+#' @param dir directory (created in function)
+#' @param latitude,longitude,split_col,d.fig inherited from parent
+#' @noRd
+create_static_markers <-
+  function(fun,
+           data = data,
+           dir = tempdir,
+           latitude = latitude,
+           longitude = longitude,
+           split_col = split_col,
+           popup = NULL,
+           label = NULL,
+           d.fig) {
+    if (is.null(popup)) {
+      data$popup <- "NA"
+      popup <- "popup"
+    }
+    if (is.null(label)) {
+      data$label <- "NA"
+      label <- "label"
+    }
+
+    # create plots
+    plots_df <-
+      data %>%
+      tidyr::drop_na(.data$conc) %>%
+      tidyr::nest(data = -dplyr::all_of(c(
+        latitude, longitude, split_col, popup, label
+      ))) %>%
+      dplyr::mutate(
+        plot = purrr::map(data, fun, .progress = "Creating Polar Markers"),
+        url = paste0(dir, "/", .data[[latitude]], "_", .data[[longitude]], "_", .data[[split_col]], ".png")
+      )
+
+    purrr::pwalk(list(plots_df[[latitude]], plots_df[[longitude]], plots_df[[split_col]], plots_df$plot),
+                 .f = ~ {
+                   grDevices::png(
+                     filename = paste0(dir, "/", ..1, "_", ..2, "_", ..3, ".png"),
+                     width = d.fig * 300,
+                     height = d.fig * 300,
+                     res = 300,
+                     bg = "transparent",
+                     type = "cairo",
+                     antialias = "none"
+                   )
+
+                   plot(..4)
+
+                   grDevices::dev.off()
+                 })
+
+    return(plots_df)
+  }
+
+#' if ggmap is not provided, have a guess
+#' @param data `plots_df` input
+#' @param ggmap,latitude,longitude,zoom inherited from parent
+#' @noRd
+estimate_ggmap <-
+  function(ggmap = ggmap,
+           data,
+           latitude = latitude,
+           longitude = longitude,
+           zoom = zoom) {
+    if (is.null(ggmap)) {
+      lat_d <- abs(diff(range(data[[latitude]])) / 2)
+      minlat <- min(data[[latitude]]) - lat_d
+      maxlat <- max(data[[latitude]]) + lat_d
+
+      lon_d <- abs(diff(range(data[[longitude]])) / 2)
+      minlon <- min(data[[longitude]]) - lon_d
+      maxlon <- max(data[[longitude]]) + lon_d
+
+      ggmap <-
+        ggmap::get_stamenmap(bbox = c(minlon, minlat, maxlon, maxlat),
+                             zoom = zoom)
+    }
+
+    return(ggmap)
+  }
+
+#' Create static map
+#' @param ggmap:facet.nrow inherited from parent
+#' @param plots_df `plots_df`
+#' @noRd
+create_static_map <-
+  function(ggmap,
+           plots_df,
+           dir,
+           latitude,
+           longitude,
+           split_col,
+           pollutant,
+           d.icon,
+           facet,
+           facet.nrow) {
+    plt <-
+      ggmap::ggmap(ggmap) +
+      ggtext::geom_richtext(
+        data = dplyr::mutate(
+          plots_df,
+          url = stringr::str_glue("<img src='{url}' width='{d.icon}'/>")
+        ),
+        ggplot2::aes(.data[[longitude]], .data[[latitude]], label = .data$url),
+        fill = NA,
+        color = NA
+      ) +
+      ggplot2::labs(x = NULL, y = NULL) +
+      theme_static()
+
+    if (length(pollutant) > 1 | !is.null(facet)) {
+      plt <-
+        plt + ggplot2::facet_wrap(ggplot2::vars(quickTextHTML(.data[[split_col]])), nrow = facet.nrow) +
+        ggplot2::theme(strip.text = ggtext::element_markdown())
+    }
+
+    return(plt)
+  }
+
+